--- f_s/src/Components/home2.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-function Home2() {
-  return (
-    <div className="carousel slide" data-ride="carousel">
-        <ol className="carousel-indicators">
-            <li></li>
-            <li></li>
-        </ol>
-        <div className="carousel-inner">
-          <div className="carousel-item active"></div>
-          <div className="carousel-item"> </div>
-        </div>
-    </div>
-  )
-}
-
-export default Home2;
-
